Restore button functionality to the recycling bin index

+++ src/pages/RecyclingBin/RecyclingBin.tsx
@@ -1,10 +1,13 @@
 import styles from "pages/RecyclingBin/RecyclingBinPage.module.scss";
 import { Link } from "react-router-dom";
-import { useState, useEffect } from "react";
+
+import { useState } from "react";
+import { useHistory } from "react-router-dom";
+
 import Button from "components/UI/Button";
 
 import { useAppSelector, useAppDispatch } from "app/hooks";
-import { selectArticlesData, deleteArticle } from "app/articlesSlice";
+import { selectArticlesData, deleteArticle, restoreArticle } from "app/articlesSlice";
 import { LoginType, selectLoginType } from "app/adminSlice";
 
 import { validPageLink, dictionaryUpdateKey } from "utils/functions";
@@ -43,6 +46,7 @@ function RecyclingBin() {
   const articlesData = useAppSelector(selectArticlesData);
   const currentLoginType = useAppSelector(selectLoginType);
   const dispatch = useAppDispatch();
+  const history = useHistory();
 
   return (
     <div className={styles.index}>
@@ -52,17 +56,37 @@ function RecyclingBin() {
 
       {
         articlesData.articles
-          .map(({ name, draftStatus }) => (
+          .map(({ name }) => (
             <>
+              {
+              /*
+                I may install material-icons since I kind of want to replace these with
+                icons... Although technically I should ask for permission first.
+               */
+              }
               <p key={name} >
                 <Button
                   styledAs="oval"
-                  onClick={
-                    function() {
-                      dispatch(deleteArticle(name));
+                  onClick={async () => {
+                    if (confirm("Do you want to restore this article?")) {
+                      await dispatch(restoreArticle(name));
+                      history.push(validPageLink(name));
                     }
-                  }
-                >Delete</Button>
+                  }}
+                >
+                  Restore
+                </Button>
+                <Button
+                  styledAs="oval"
+                  onClick={() => {
+                    if (confirm("Permenantly delete this article?")) {
+                      dispatch(deleteArticle(name));
+                      history.push("/");
+                    }
+                  }}
+                >
+                  Delete
+                </Button>
                 <Link to={"/admin/recycling_bin/" + validPageLink(name)}>
                   {name}
                   <DaysLeft value={15}/>
